deserializer

+++ src/Shapes/SerializableEllipse.java
@@ -36,9 +36,9 @@ public class SerializableEllipse extends SerializableShape {
 
 		Rectangle2D bounds = ellipse2D.getBounds();
 
-		sb.append(bounds.getX());
+		sb.append((int)bounds.getX());
 		sb.append(' ');
-		sb.append(bounds.getY());
+		sb.append((int)bounds.getY());
 		sb.append(' ');
 		sb.append(getWidth());
 		sb.append(' ');
